Moved the java files into the typical directory structure of a java app.
