+++ common/ECSResourcesCheck/src/main/java/Comparer.java
