+++ common/ECSResourcesCheck/src/main/resources/META-INF/MANIFEST.MF
