+++ common/ECSResourcesCheck/src/main/resources/base.json
